backend/requirements.txt
# Web Framework
fastapi>=0.135.1

# Database Toolkits
sqlalchemy~=2.0.48
psycopg2-binary~=2.9.11

# Utilities (Always gets the latest)
debugpy
python-dotenv
uvicorn
gunicorn
flake8
pytest
black
pytest-asyncio
httpx